Refs #7, change authentication method to use cookies.
The login.php script now sets two cookies, one for the user_id and the
other for the auth_token. These cookies are used for authentication on
any subsequent scripts. Cookies by default expire after a week; this can
be changed in config.php. The auth_username and auth_password fields are
only required for login.php, but would be ignored by any other scripts.