Refs #30, adds images to posts and changes config for media files.
Instead of having an AVATAR_ABSOLUTE_PATH and RIFF_ABSOLUTE_PATH, there
is now a MEDIA_ABSOLUTE_PATH which will have three subfolders: avatars,
posts, and riffs. More may be added as features are added. There is also
an IMAGE_URL variable now, which is expected to be something like
"https://yoursite.com/media". This is expected to be the web-facing
entry point for the MEDIA_ABSOLUTE_PATH on the server.