Refs #27, adds more search functions for users.
get-users-nearby has been renamed to search-users-nearby and now accepts
an array of tags that it will filter by. search-users-trending is a new
script that accepts tags that it will filter by as well as a "time"
parameter, which can be "day", "week", "month", or "all", and it counts
upvotes only within that time span to get the current most popular
users.