Refs #33, adds get-messages script.
get-messages returns an array of users that are involved in the
conversation and an array of messages. It accepts "page" and "limit"
pagination parameters, and requires an "id" that is a valid user id. If
there are no messages found, it will still return successful but with an
empty array.