Refs #51, adds leave-conversation and removes get-conversation.
The get-conversation script was used to get a conversation between the current user and exactly one other user, but this didn't always make sense because there could be conversations with more than 2 people that shrunk down to 2 people. In these cases the system can't tell which one is supposed to be the direct channel from one user to another so it can't make that distinction.

In the future certain conversations might be marked as the primary conversation between two users but for now all conversations are potentially multi-way conversations, and a private chat is just a special case of two-way conversation. There is no way to add members to a conversation after it has been created, although you can remove yourself.