Refs #55, creates an uninstall script for the database.
ryff.sql has been moved to resources/db/install.sql, and is accompanied by resources/db/uninstall.sql. The create database and use  syntax are no longer in the install script, so that it can be used for either the main or test database.

connect-db.php no longer includes config.php, that is done in global.php. connect-db will connect to the test database instead if TEST_MODE is defined and truthy.

A lot of documentation has been added to SQL scripts, the config file, the
global file, and the connect database file. The formatting on the foreign key constraints has also been changed in install.sql.