Refs #55, adds get_unique_word() and moves get_test_user(), get_test_post() to TestEnvironment.
get_unique_word() will be used for usernames and emails, since they have to be unique. get_test_post() has been reworked to take an array of tags and an array of usernames to mention. The memory limit has been set to 1024MB in test-environment.class.php, because initializing all of the words takes up a lot of memory.