Refs #55, adds test media and sample media folders.
The test media folder is set as another variable and is used globally when TEST_MODE is set to true in config.php. The sample media folder should be filled by the user, right now it just has one empty file of each type. This will be the pool of files used when running the populate test script and other test scripts.