Refs #68, adds an autoloader and renames model class files.
The .class.php files that the models are stored in are now named the same
thing as the class, so the autoloader can easily include them. The test files
are included only if in test mode. All of the require_once(*.class.php)
calls have been removed. The autoloader does not include the files in
resources/tests/api_tests/, because those files are included separately only
by the ApiTests class, and in a separate autoloader-esque fashion.