Reduces memory usage of TestEnvironment, makes words.txt smaller.
Running tests now use <30M, where they used to use ~220M. This was almost entirely due to the word banks. I have reduced the word bank size to 100,000 words, and the words are in the correct format in the text file already instead of filtering, setting to lowercase, removing nonalpha characters, etc.

Also fixes an error where if get_unique_word() ran out of words, it would get stuck in an infinite loop trying to get a unique hash.