Refs #78, adds audio conversion when adding posts.
Audio is converted using ffmpeg to 128k and 256k AAC-encoded .m4a files. Post objects have both riff_url and riff_hq_url attributes now. If a post fails to upload, it might be because ffmpeg is not installed, or failed to convert the file (either because it was an invalid file or for some other reason). There is currently no way to tell which of these is the reason.

Conversion is done with exec(), and it blocks, for now. Audio conversion is not currently done in the background, but it may be in the future.