Refs #78, adds delayed audio conversion.
The duration of posts is no longer required and will be ignored. The duration will be gleaned from ffprobe or avprobe. If the codec is AAC, the file will be immediately available for listening and a post will be returned from add-post. If The codec is something else but it is convertible, the post will not yet be available for listening.

A service that has yet to be written will convert the audio in the background and mark the post as active when it has done, and will notify the user of success or failure (also yet to be written).